py/requirements.txt
grpcio-tools ~= 1.30
numpy ~= 1.20
redis ~= 5.0.1